+++ processor/lib/processor_web/error_json.ex
@@ -1,5 +1,0 @@
-defmodule ProcessorWeb.ErrorJSON do
-  def render(template, _assigns) do
-    %{errors: %{detail: Phoenix.Controller.status_message_from_template(template)}}
-  end
-end